+++ .gitignore
@@ -11,7 +11,6 @@ dist
 build
 eggs
 parts
-bin
 var
 sdist
 develop-eggs
