+++ README.md
@@ -1,19 +1,27 @@
 # Matching Software for Speed Friending and Dating Events
 [![Build Status](https://travis-ci.org/DiffSK/configobj.svg?branch=master)](https://travis-ci.org/machinekoder/speed-friending-and-dating-matcher)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/machinekoder/speed-friending-matcher/blob/master/LICENSE)
+[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/ambv/black)
 
 This application is designed to make your life as organizer of speed friending or speed dating events easier. I created this software for a [local speed friending event in Vienna, Austria](https://www.meetup.com/de-DE/speed-friending-events/) to give back value to the event organizers. For me, the project additionally serves as a playground for software engineering best practices. The application was implemented in an agile, test-driven development process applying all development best practices so far known to me.
 
 ## Installing the application
-Install the application dependencies using pip:
+To install the live coding environment run:
+
 ```bash
-sudo pip install -r requirements.txt
+python setup.py install
+```
+
+or install it via pip
+
+```bash
+pip install speed-friending-matcher
 ```
 
 ## How to use it
 Run the speed-friending-matcher from the command line:
 ```
-usage: speed-friending-matcher.py [-h] -i INPUT -o OUTPUT [-m MATCHMAKER]
+usage: speed_friending_matcher [-h] -i INPUT -o OUTPUT [-m MATCHMAKER]
 
 Matchmaking application for speed friending events
 
@@ -29,7 +37,7 @@ optional arguments:
 
 For example:
 ```bash
-./speed-friending-matcher.py -i csv:example/sample.csv -o todo:test.txt
+speed_friending_matcher -i csv:example/sample.csv -o todo:test.txt
 ```
 
 ## Importer Plugins
